[Refactor] #166 - 스와이프 뒤로가기 복구를 WSSComponent로 통합 (delegate 소유를 네비게이션 컨트롤러로 전환)
NovelDetailFeature·LibraryFeature에 각각 복제돼 있던 SwipeBackEnabler를
WSSComponent의 `.enableSwipeBack()` 하나로 합친다. #166에서 한쪽만 고쳐져
갈라진 상태였고, 그 탓에 작품 상세를 다녀오면 서재 쪽 반납 가드까지
무력화돼 루트에서 내비게이션이 얼어붙을 수 있었다.

합치면서 delegate 소유 방식을 바꿨다. 기존에는 제스처 delegate를 Coordinator가
가로챘다가 화면을 떠날 때 반납했는데, delegate가 약한 참조라 반납 타이밍을
정확히 맞춰야 했고(늦게 도착한 async 블록이 반납을 되돌리는 경합까지 있었다)
복제본 중 하나가 반납을 빠뜨리면 다른 화면의 반납까지 무너졌다.
delegate를 UINavigationController 자신에게 맡기면 그 수명이 스택과 같아져
반납·경합 가드·복제본 동기화가 모두 불필요해진다.

대신 UIKit 기본 delegate가 보던 조건을 직접 챙긴다 — 루트 여부,
전환 진행 중 여부, 그리고 뒤로가기 버튼 숨김 여부. 마지막 조건은
NovelReviewView가 "닫기 전 확인 알럿"을 그 성질에 기대고 있어 필요하다.
